RabbitMQ works #42
