+++ plot-weather-microservice/requirements.txt
@@ -5,3 +5,4 @@ matplotlib==3.5.1
 scipy==1.7.3
 numpy==1.22.2
 pandas==1.4.0
+pika==1.2.0
